requirements-local.txt
# Core dependencies
pydantic>=2.0.0
pydantic-ai>=0.1.1
python-dotenv>=1.0.0

# Database dependencies
psycopg2-binary>=2.9.9

# HTTP and networking
httpx>=0.25.0
httpcore>=1.0.0
anyio>=4.0.0

# MCP dependencies are installed via npm/npx in the Dockerfile
mcp==1.6.0

# Utilities
typer>=0.9.0
rich>=13.6.0
logfire>=0.9.0

# API Server
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Authentication
pyjwt>=2.8.0
